# cais/src/main/java/br/com/compass/cais/services/dto/request/ShipRequestDTO.java
package br.com.compass.cais.services.dto.request;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import javax.validation.constraints.NotBlank;
import javax.validation.constraints.NotNull;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class ShipRequestDTO {

    @NotBlank
    private String name;

    @NotNull
    private Double weight;

}